// frontend/src/shared/api/model/placeResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Tarikma API
 * Backend API for Tarik.ma
 * OpenAPI spec version: v1
 */

export interface PlaceResponse {
  id?: number;
  cityId?: number;
  cityName?: string;
  name?: string;
  description?: string;
  image?: string;
  address?: string;
  category?: string;
  latitude?: number;
  longitude?: number;
  mapPlaceId?: string;
  createdAt?: string;
  createdBy?: number;
}